Separate InputClassic components
