+++ lib/components/Input/CustomHelperText.d.ts
@@ -1,0 +1,4 @@
+import { FC } from 'react';
+import { InputProps } from './InputClassic';
+declare const CustomHelperText: FC<Pick<InputProps, 'helperText' | 'hasCounter' | 'maxLength' | 'value' | 'success'>>;
+export default CustomHelperText;
