+++ lib/components/Input/CustomInput.d.ts
@@ -1,0 +1,4 @@
+import { FC } from 'react';
+import { InputProps } from './InputClassic';
+declare const CustomInput: FC<Pick<InputProps, 'value' | 'placeholder' | 'onChange' | 'inputRef' | 'maxLength' | 'success'>>;
+export default CustomInput;
