+++ lib/components/Input/CustomLabel.d.ts
@@ -1,0 +1,4 @@
+import { FC } from 'react';
+import { InputProps } from './InputClassic';
+declare const CustomLabel: FC<Pick<InputProps, 'label' | 'success'>>;
+export default CustomLabel;
