+++ src/huexports.ts
@@ -10,5 +10,5 @@ export { default as HuPills } from './components/Pills/Pills';
 export { default as HuSkeleton } from './components/Skeleton/Skeleton';
 export { default as HuTitle } from './components/Title/Title';
 export { default as HuTabs } from './components/Tabs/Tabs';
-export { default as HuInputClassic } from './components/InputClassic/InputClassic';
-export { default as HuFormInputClassic } from './components/InputClassic/FormInputClassic';
+export { default as HuInputClassic } from './components/Input/InputClassic';
+export { default as HuFormInputClassic } from './components/Input/FormInputClassic';
